==> Extraction_Modelling/0. Full Dataset Merge.ipynb <==
# Full Dataset Merge

Create a data set contain 854 cases bases on Lilian's PDF parse

## Json include all the text

import json
full_data = json.load(open('amy_cases.json', 'r'))

keys = list(full_data.keys())

keys[1:5]

len(keys)

import re
file_name = []
for key in keys:
    key = key.replace('/Users/liyannie/Dropbox/Westlaw_cases/', '').replace('.pdf', '')
    key = re.sub("\d", '', key)
    key = key.replace('-', '').lstrip()
    file_name.append(key)

len(file_name)

file_text = []
for key in keys:
    file_text.append(full_data[key])

len(file_text)

file_text[500][0:100]

file_name[500]

## Store in Json

json.dump({'file_name': file_name,
           'raw_text': file_text},
          open('full_name_text.json', 'w'))

## Load Created Json File

import json
data = json.load(open('full_name_text.json', 'r'))


data.keys()

file_name = data['file_name']
raw_text = data['raw_text']

len(file_name)

len(raw_text)
