==> Extraction_Modelling/1. Edge and Node List.ipynb <==
# Create Edge and  Node List

1. Extract Court
2. Create a edge list of the citation network
3. Create a node list contains the case code, name and court

### Read Case text data

import json
data = json.load(open('full_name_text.json', 'r'))
file_name = data['file_name']
raw_text = data['raw_text']

len(file_name)

len(raw_text)

## 1. Extract Court

from lexnlp.extract.en.dict_entities import get_entity_name, entity_config
import lexnlp.extract.en.courts
import pandas

court_df = pandas.read_csv("https://raw.githubusercontent.com/LexPredict/lexpredict-legal-dictionary/1.0.5/en/legal/us_courts.csv")

# Create config objects
court_config_data = []
for _, row in court_df.iterrows():
    c = entity_config(row["Court ID"],row["Court Name"],row["Court Type"])
    court_config_data.append(c)


def get_court(text):
    text_court =lexnlp.extract.en.courts.get_courts(text, court_config_data)
    court_in_pdf = []   
    for entity in text_court:
        court_in_pdf.append(entity[0][1:3])
    try:
        return court_in_pdf[1]
    except:
        return (None, None)   

court_list = []
court_type = []
for raw in raw_text:   
    court = get_court(raw)[0]
    court_list.append(court)
    ctype = get_court(raw)[1]
    court_type.append(ctype)


len(court_list)

len(court_type)

court_na = []
for i in range(len(file_name)):
    if court_list[i] == None: 
        court_na.append(i)       

len(court_na)

## 2. Create a edge list of the citation network

import re
# import a dictionary contains court information
from reporters_db import EDITIONS
for key in list(EDITIONS.keys()):
    EDITIONS[key.replace(" ", "")]=EDITIONS[key].replace(" ", "")
    
# regular expression to describe citation pattern
CITATION_PTN = r"""
(?:[\s,:\(]|^)
(
(\d+)\s+
({reporters})(\s|[a-z])+
(\d+)
)
""".format(reporters='|'.join([re.escape(i) for i in EDITIONS]))
CITATION_PTN_RE = re.compile(CITATION_PTN, re.IGNORECASE | re.MULTILINE | re.DOTALL | re.VERBOSE)


def get_citations(raw_text):
    result = CITATION_PTN_RE.findall(raw_text)
    citations = []
    for cite in result:
        the_cite = cite[0].replace("at ", "").replace("  "," ").replace("  "," ").replace("\n", " ")
        if len(the_cite)<=30:
            citations.append(the_cite)
    # remove duplicated citations
    citations = list(dict.fromkeys(citations))
    return citations

len(raw_text)

len(file_name)

import pandas as pd

cite_dict = {}
for i in range(len(file_name)):
    cite_dict[file_name[i]] = get_citations(raw_text[i])

# A list contain the case number of the PDF file
case_number = []
for i in range(len(file_name)):
    # Case number is the first citation number in the citation list
    case_number.append(cite_dict[file_name[i]][0])

cite_number_dict = {}
for i in range(len(file_name)):
    cite_number_dict[case_number[i]] = cite_dict[file_name[i]]

for key in case_number:
    cite_number_dict[key] = cite_number_dict[key][1:]

len(case_number)

len(cite_number_dict)

cite_df = pd.DataFrame.from_dict(cite_number_dict, orient='index')

cite_df['file']= cite_df.index

cite_df.head()

citation_df= pd.melt(cite_df,id_vars=['file']). \
            drop(['variable'], axis=1).dropna(). \
            rename(index=str, columns={"file": "from",
                                      "value":"to"})

citation_df.to_csv("cite_egde.csv", sep=',', encoding='utf-8', index=False)

## 3. Create a node list contains the case code, name and court

Nodes = pd.DataFrame(
    {'case_code': case_number,
     'case_name': file_name,
     'court_from': court_list,
     'court_type':court_type,
    })


Nodes.head()

Nodes.to_csv("cite_node.csv", sep='\t', encoding='utf-8',index=False)
