==> Extraction_Modelling/2. Full Extractions.ipynb <==
# Addtional Extraction for the Node List

1. Case Result, 
2. Author of the opinion text
3. Judge Panel
4. Opinion text


### Read text

import json
pdf2text = json.load(open("full_name_text.json", 'r'))

pdf2text.keys()

file_name = pdf2text['file_name']

raw_text = pdf2text['raw_text']

## 1. Case Result

import re

for layer in raw_text[10].split("\n"):
    if re.search(r'(AFFIRMED|APPEAL|VACATED|REVERSED|DISMISSED|AMENDED)', layer) != None:
        print(layer)

decision = [None]*len(file_name)
for i in range(len(file_name)):
    for layer in raw_text[i].split("\n"):
        if re.search('^(AFFIRMED|APPEAL|VACATED|REVERSED|DISMISSED|AMENDED)', layer) != None:
            decision[i] = layer
    

de_no = []
for i in range(len(decision)):
    if decision[i] == None:
        de_no.append(i) 

len(de_no)

for i in range(len(file_name)):
    if decision[i] == None:
        for layer in raw_text[i].split("\n"):
            if re.search('(en banc.*denied\.|MOTION|petition.*denied\.|Remanded|Affirmed|reversed\.|Vacated and remanded|Petition granted)', layer) != None:
                decision[i] = layer

de_no = []
for i in range(len(decision)):
    if decision[i] == None:
        de_no.append(i) 

len(de_no)

## 2. Author of the Opinion Text

for layer in raw_text[9].split("\n"):
    if len(layer) <40 and \
    re.search(r'^(?!Before)(?!dissenting-in-part).*Judge.$', layer) != None:
        print(layer.split(",")[0])

author = [None]*len(file_name)
for i in range(len(file_name)):
    for layer in raw_text[i].split("\n"):
        if len(layer) <40 and \
        re.search(r'^(?!Before)(?!dissenting-in-part).*Judge.$', layer) != None:
            author[i] = layer.split(",")[0]

author_judge = author
for i in range(len(file_name)):
    if author[i]!=None and len(author[i].split("  "))>1:
        author_judge[i] = author[i].split("  ")[-1]
    
    

## 3. Judge Panel

judge_panel = [None]*len(file_name)
for i in range(len(file_name)):
        for layer in raw_text[i].split("\n"):
            if re.search(r'^Before(.*)Judge(.*)', layer) != None:
                    judge_panel[i] = layer

## 4. Opinion Text

### Identify the start point of the opinion text

op_start_layer = [None]*len(file_name)
for i in range(len(file_name)):
    for layer in raw_text[i].split("\n"):
        if len(layer) <40 and \
        re.search(r'^(?!Before)(?!dissenting-in-part).*Judge.$', layer) != None:
            op_start_layer[i] = layer

start_index = [None]*len(file_name)
for i in range(len(file_name)):
    if op_start_layer[i] != None:
        start_index[i] = (raw_text[i].split("\n").index(op_start_layer[i]))+1

### Identify the end point of the opinion text

op_end_layer = [None]*len(file_name)
for i in range(len(file_name)):
    for layer in raw_text[i].split("\n"):
        if re.search('^(AFFIRMED|APPEAL|VACATED|REVERSED|DISMISSED|AMENDED)', layer) != None:
            op_end_layer[i] = layer

end_index = [None]*len(file_name)
for i in range(len(file_name)):
    if op_end_layer[i] != None:
        end_index[i] = (raw_text[i].split("\n").index(op_end_layer[i]))

### Extract the opinion text based on the anchor 

opinion_text = [None]*len(file_name)

for i in range(len(file_name)):
    if start_index[i] != None and end_index[i]!= None:
        opinion_text[i]=''.join(raw_text[i].split("\n")[start_index[i]:end_index[i]])

op_no = []
for i in range(len(decision)):
    if opinion_text[i] == None:
        op_no.append(i) 

len(op_no)

## Store the Extractions as Meta Data

### Append to the node list `.csv`

import pandas as pd

meta = pd.read_csv("cite_node.csv", sep="\t")

meta.head()

full_data = meta.assign(result = decision,
                  author = author_judge,
                  judge_panel = judge_panel)

full_data.to_csv('extraction854.csv',index=False, sep='\t')

### Store `author`, `result`, `opinion text`  to json

json.dump({'author': author_judge,
           'decision': decision,
           'opinion':opinion_text},open('decision_text.json','w'))

import json

op_data = json.load(open('decision_text.json','r'))

op_data.keys()

len(op_data['opinion'])
