==> Extraction_Modelling/3. Clean Opinion Text.ipynb <==
# Opinion Text Clean

**Remove references, special characters, Thomson Reuters marks in the opinion text**

### Load data

import json
full_data = json.load(open('decision_text.json', 'r'))

full_data.keys()

op_text = full_data['opinion']

op_text[0][0:100]

### Identify references

import re
from reporters_db import EDITIONS
for key in list(EDITIONS.keys()):
    EDITIONS[key.replace(" ", "")]=EDITIONS[key].replace(" ", "")

# regular expression to describe citation pattern

CITATION_PTN = r"""
(?:[\s,:\(]|^)
(
(\d+)\s+
({reporters})(\s|[a-z])+
(\d+)
)
""".format(reporters='|'.join([re.escape(i) for i in EDITIONS]))
CITATION_PTN_RE = re.compile(CITATION_PTN, re.IGNORECASE | re.MULTILINE | re.DOTALL | re.VERBOSE)

# regular expression to describe regulation pattern
REGULATION_PTN = r"""
(
(\d+)\s*
(U\.?S\.?C\.?|C\.?F\.?R\.?)\s*
(Sec(?:tion|\.)?|§)?\s*
(\d+[\da-zA-Z\-]*)
)"""
REGULATION_PTN_RE = re.compile(REGULATION_PTN, 
                               re.IGNORECASE | re.MULTILINE | re.DOTALL | re.VERBOSE)

PUBLIC_LAW_PTN = r"""
(
Pub(?:lic|\.)\s+L(?:aw|\.)(?:\s+No.?)?\s+\d+\-\d+
|
\d+\s+Stat\.\s+[\d-]+
)
"""
PUBLIC_LAW_PTN_RE = re.compile(PUBLIC_LAW_PTN, 
                               re.IGNORECASE | re.MULTILINE | re.DOTALL | re.VERBOSE)
PUBLIC_LAW_SUB_RE = re.compile(r'.+?(\d+\-\d+)',
                               re.MULTILINE | re.DOTALL)




FED_CIR_RE =  re.compile(r'\(Fed.Cir.\d*\)', re.MULTILINE | re.DOTALL)

MARK_RE = re.compile(r'©\s\d+\sThomson Reuters.', re.MULTILINE | re.DOTALL)

### Write a function to clean text
Remove references, special characters, Thomson Reuters marks in the opinion text

replacements = [
    (CITATION_PTN_RE, ''), # remove other case reference
    (REGULATION_PTN_RE, ''), # remove regulation reference
    (PUBLIC_LAW_PTN_RE, ''), # remove public law reference
    (FED_CIR_RE, ''), # remove Fed.Cir
    (MARK_RE, ''), # remove Thomson Reuters mark
    (r'[^\w]', ' '), # remove special characters
    (r'\d+', ' '), # remove numbers
    (r'\s+[A-Za-z]{1,2}\s+', ' '), # remove words with only one or two letters
    (r'\s+', ' '), # remove consecutive space
    (r'\s+[A-Za-z]{1,2}\s+', ' '), # remove words with only one or two letters
    (r'\s+col\s+', ' '), # remove 'col' references
    (r'\s+[A-Za-z]{1,2}\s+', ' ') # remove words with only one or two letters
]

def clean_citation(input_str):
    for old, new in replacements:
        input_str = re.sub(old, new, input_str)
    return input_str

#### Test the `clean_citation` function

# Origional Text
op_text[5][500:2100]

# Cleaned Text
clean_citation(op_text[5][500:2100])

## Use `clean_citation` to clean all opinion text

cleaned_text = [None]*len(op_text)

for i in range(len(op_text)):
    if op_text[i] != None:
        cleaned_text[i]= clean_citation(op_text[i])

# Write to json file

full_data.update({'cleaned_text':cleaned_text})

full_data.keys()

json.dump(full_data, open('decision_text.json', 'w'))

# Write to csv

import pandas as pd

text_import = pd.DataFrame(
    {'cleaned_text': cleaned_text
    })


text_import.head()

text_import.to_csv('cleaned_text.csv',sep='\t')
