==> Extraction_Modelling/4.2 Citation Between Nodes.Rmd <==
---
title: "Citation Between Available Cases"
author: "Yi Yin"
date: "4/27/2019"
output:
  pdf_document: default
  html_document: default
editor_options:
  chunk_output_type: console
---

```{r Setup, include=FALSE, results='hide', warning=FALSE}
library(knitr)
opts_chunk$set(fig.path="figures/between_nodes/",
               cache.path="cache/between_nodes/",
               cache=TRUE,
               echo=TRUE,
               message=FALSE,
               warning=FALSE)  
```  


```{r preamble, message=FALSE, warning=FALSE, include=FALSE}
rm(list = ls())
library(tidyverse)
```


```{r edge list, message=FALSE}
links <-read.csv("./cite_egde.csv", sep = ",", 
                 colClasses = 'character')%>%
        filter(to %in% from)

edges<-links%>%group_by(to)%>%summarise(freq=n())%>%
        arrange(desc(freq))%>%filter(freq>9)
links<- links%>%filter(to %in% edges$to)
```


```{r network graph, fig.align='center', fig.width=12, fig.height=12}
library("igraph")
net <- graph_from_data_frame(d=links, directed=T)


vex <-tibble(node = rep(NA, length(V(net)$name)))%>%
        mutate(node = V(net)$name)
vex$indegree <- degree(net, mode = "in",normalized = T)

bar<-vex%>%arrange(desc(indegree))%>%
        .[25,'indegree']%>%
        unlist(use.names = FALSE)

vex <-vex%>%mutate(label = ifelse(indegree >=bar, 
                                  node, NA))

vex <-vex%>%mutate(ncolor = ifelse(indegree >=bar, 
                                   "#923740", '#68B3D0'))

V(net)$color <-vex$ncolor
V(net)$label <-NA
V(net)$label.family <-"Palatino"
V(net)$label.color <-"#020909"
V(net)$label.font <- 2
V(net)$label.cex <- 0.5
V(net)$label.degree<- -pi/3
V(net)$label.dist <-0.1
V(net)$size <- vex$indegree*60

# V(net)$vertex.frame.color<-"#2C2D25"
# V(net)$vertex.frame.width<-0.0001
# V(net)$vertex.shape<-"circle"

E(net)$width <- 0.6
E(net)$edge.color <- "#2C2D25"
E(net)$arrow.size <- 0.35


plot(net, 
     layout=layout_with_kk(net),
     vertex.frame.width=0.0001,
     vertex.shape="circle",
     rescale=TRUE)

```


```{r, include = FALSE, eval=FALSE}

plot(all, layout=All_layout, vertex.size=2, vertex.label=V(All)$name,
vertex.color="green", vertex.frame.color="red", edge.color="grey",  
edge.arrow.size=0.01, 
rescale=TRUE,
vertex.label=NA,
vertex.label.dist=0.0,
vertex.label.cex=0.5, 
add=FALSE, 
vertex.label.font=.001)

```
